Get Coding SSMs

Convenience function for loading coding Simple Somatic Mutations (SSM) from the bundled data GAMBLR.data::sample_data.

Usage

get_coding_ssm(
  these_sample_ids = NULL,
  these_samples_metadata = NULL,
  projection = "grch37",
  this_seq_type = "genome",
  tool_name = "slms-3",
  min_read_support = 3,
  include_silent = TRUE,
  verbose = FALSE,
  ...
)

Arguments

these_sample_ids

Optional, a vector of multiple sample_id (or a single sample ID as a string) that you want results for.

these_samples_metadata

Optional, a metadata table (with sample IDs in a column) to subset the return to. If not provided (and if these_sample_ids is not provided), the function will return all samples from the specified seq_type in the metadata.

projection

Reference genome build for the coordinates in the MAF file. The default is grch37.

this_seq_type

The this_seq_type you want back, default is genome.

tool_name

Optionally specify which tool to report variant from. The default is slms-3, also supports "publication" to return the exact variants as reported in the original papers.

min_read_support

Only returns variants with at least this many reads in t_alt_count.

include_silent

Logical parameter indicating whether to include silent mutations into coding mutations. Default is TRUE.

verbose

Set to FALSE to minimize the output to console. Default is TRUE. This parameter also dictates the verbosity of any helper function internally called inside the main function.

...

Any additional parameters.

Value

data frame

Details

This "bare bones" function was developed to retrieve coding SSM calls for non-GSC-users. Effectively retrieve coding SSM calls. Multiple filtering parameters are available for this function. For more information on how to implement the filtering parameters, refer to the parameter descriptions as well as examples in the vignettes. This function depends on the bundled sample data in this package.

Examples

 # Get mutations from exome data originally aligned to grch37
ssm_exomes_grch37 = get_coding_ssm(projection = "grch37",this_seq_type = "capture")
#> Using the bundled SSM calls (.maf) calls in GAMBLR.data...
#> Using the bundled metadata in GAMBLR.data...
#> after linking with metadata, we have mutations from 1787 samples

# Get mutations from genome data, hg38 build
ssm_genomes_hg38 = get_coding_ssm(projection = "hg38",this_seq_type = "genome")
#> Using the bundled SSM calls (.maf) calls in GAMBLR.data...
#> Using the bundled metadata in GAMBLR.data...
#> after linking with metadata, we have mutations from 860 samples